requirements.txt
google-generativeai
chromadb
sentence-transformers
langchain
python-docx
